# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Past Life Archive API
  version: 0.2.5
  description: PRD 2.5 assessment, result presentation, unlock, and share-asset API.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /auth/archive/{resultId}:
    get:
      operationId: getArchiveDetail
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ResultId' }]
      responses:
        '200':
          description: Full basic result for an archive entry owned by the signed-in user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ArchiveDetail' }
        '401': { description: Authentication required }
        '403': { description: Archive entry is not owned by the signed-in user }
    delete:
      operationId: deleteArchiveEntry
      description: Removes only the authenticated user's archive association. The anonymous result and other users' archive entries are retained. Repeated deletion is idempotent.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ResultId' }]
      responses:
        '204': { description: Archive association is absent after the request }
        '401': { description: Authentication required }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ResultId: { name: resultId, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    ViewMode:
      type: string
      enum: [VIDEO, TEXT]
...
    BasicResult:
      type: object
      required: [resultId, recordNo, headline, character, lifeSummary, highlights, image, blocks, disclaimer]
      properties:
        resultId: { type: string }
        recordNo: { type: integer, minimum: 1, maximum: 99 }
        headline: { type: string }
        character:
          type: object
          description: Fictional character dossier; it does not infer the user's real identity.
        lifeSummary:
          type: object
          required: [ageAtDeath, lifespanLabel, endingTitle]
          properties:
            ageAtDeath: { type: integer, minimum: 1 }
            lifespanLabel: { type: string }
            endingTitle: { type: string }
        highlights: { type: array, minItems: 8, maxItems: 8, items: { type: object } }
        image: { type: object }
        blocks: { type: array, minItems: 6, maxItems: 6, items: { type: object } }
        disclaimer: { type: string }
    ArchiveDetail:
      allOf:
        - { $ref: '#/components/schemas/BasicResult' }
        - type: object
          required: [viewMode]
          properties:
            viewMode: { $ref: '#/components/schemas/ViewMode' }
...